+++ grpc_servicer/smg_grpc_servicer/tokenspeed/server.py
@@ -26,7 +26,7 @@ from grpc_health.v1 import health_pb2_grpc
 from grpc_reflection.v1alpha import reflection
 from smg_grpc_proto import tokenspeed_scheduler_pb2_grpc
 from smg_grpc_proto.generated import tokenspeed_scheduler_pb2
-from tokenspeed.runtime.server_args import ServerArgs
+from tokenspeed.runtime.utils.server_args import ServerArgs
 
 from smg_grpc_servicer.tokenspeed.health_servicer import TokenSpeedHealthServicer
 from smg_grpc_servicer.tokenspeed.scheduler_launcher import launch_engine
